Refactor graph controls into unified control list
