+++ packages/doenetml-worker-javascript/src/components/Graph.js
@@ -11,109 +11,14 @@ import {
     returnRoundingAttributes,
     returnRoundingStateVariableDefinitions,
 } from "../utils/rounding";
+import {
+    GRAPH_CONTROL_COMPONENT_TYPES,
+    GRAPH_CONTROL_DESCENDANT_CONFIGS,
+    GRAPH_CONTROL_VARIABLE_NAMES,
+} from "../utils/graphControls";
 // PreFigure conversion architecture and extension guide:
 // see src/utils/prefigure/README.md
 import { returnGraphPrefigureStateVariableDefinitions } from "../utils/prefigure/stateVariable";
-
-/**
- * Extracts display/rounding metadata shared by all graph control payloads.
- */
-function extractControlDisplaySettings(stateValues) {
-    return {
-        label: typeof stateValues.label === "string" ? stateValues.label : "",
-        labelHasLatex: Boolean(stateValues.labelHasLatex),
-        displayDigits: stateValues.displayDigits,
-        displayDecimals: stateValues.displayDecimals,
-        displaySmallAsZero: stateValues.displaySmallAsZero,
-        padZeros: stateValues.padZeros,
-    };
-}
-
-/**
- * Returns true when vertices contain at least `minVertexCount` 2D finite points.
- */
-function hasValid2DNumericalVertices(numericalVertices, minVertexCount) {
-    return (
-        Array.isArray(numericalVertices) &&
-        numericalVertices.length >= minVertexCount &&
-        numericalVertices.every(
-            (vertex) =>
-                Array.isArray(vertex) &&
-                vertex.length >= 2 &&
-                Number.isFinite(Number(vertex[0])) &&
-                Number.isFinite(Number(vertex[1])),
-        )
-    );
-}
-
-/**
- * Computes the arithmetic center of 2D numerical vertices.
- * Returns null when the computed center is non-finite.
- */
-function calculate2DVerticesCenter(numericalVertices) {
-    const center = numericalVertices.reduce(
-        (acc, vertex) => {
-            acc[0] += Number(vertex[0]);
-            acc[1] += Number(vertex[1]);
-            return acc;
-        },
-        [0, 0],
-    );
-
-    center[0] /= numericalVertices.length;
-    center[1] /= numericalVertices.length;
-
-    if (!Number.isFinite(center[0]) || !Number.isFinite(center[1])) {
-        return null;
-    }
-
-    return {
-        x: center[0],
-        y: center[1],
-    };
-}
-
-/**
- * Resolves effective addControls mode for shapes with independent center and size draggability.
- *
- * Returns:
- * - `centerMode` when only center movement is allowed
- * - `sizeMode` when only size movement is allowed
- * - `compositeMode` when both are allowed
- * - `null` when neither is allowed
- */
-function resolveCompositeControlsMode({
-    requestedMode,
-    draggable,
-    verticesDraggable,
-    centerMode,
-    sizeMode,
-    compositeMode,
-}) {
-    if (requestedMode === centerMode) {
-        return draggable ? centerMode : null;
-    }
-
-    if (requestedMode === sizeMode) {
-        return verticesDraggable ? sizeMode : null;
-    }
-
-    if (requestedMode !== compositeMode) {
-        return requestedMode;
-    }
-
-    if (draggable && !verticesDraggable) {
-        return centerMode;
-    }
-    if (verticesDraggable && !draggable) {
-        return sizeMode;
-    }
-    if (!draggable && !verticesDraggable) {
-        return null;
-    }
-
-    return compositeMode;
-}
 
 export default class Graph extends BlockComponent {
     constructor(args) {
@@ -671,844 +576,74 @@ export default class Graph extends BlockComponent {
             },
         };
 
-        stateVariableDefinitions.draggablePointsForControls = {
+        stateVariableDefinitions.graphicalDescendantsForControls = {
             forRenderer: true,
             returnDependencies: () => ({
                 addControls: {
                     dependencyType: "stateVariable",
                     variableName: "addControls",
                 },
-                pointDescendants: {
+                controlDescendants: {
                     dependencyType: "descendant",
-                    componentTypes: ["point"],
-                    variableNames: [
-                        "numericalXs",
-                        "draggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
+                    componentTypes: GRAPH_CONTROL_COMPONENT_TYPES,
+                    variableNames: GRAPH_CONTROL_VARIABLE_NAMES,
+                    variablesOptional: true,
                 },
             }),
-            definition({ dependencyValues }) {
+            definition({ dependencyValues, componentInfoObjects }) {
                 if (dependencyValues.addControls === "none") {
                     return {
                         setValue: {
-                            draggablePointsForControls: [],
+                            graphicalDescendantsForControls: [],
                         },
                     };
                 }
 
-                const draggablePointsForControls = [];
+                const graphicalDescendantsForControls = [];
+                const countersByType = Object.fromEntries(
+                    GRAPH_CONTROL_DESCENDANT_CONFIGS.map((config) => [
+                        config.controlType,
+                        0,
+                    ]),
+                );
 
-                for (const [pointInd, pointDescendant] of (
-                    dependencyValues.pointDescendants ?? []
-                ).entries()) {
-                    const stateValues = pointDescendant.stateValues ?? {};
-                    const numericalXs = stateValues.numericalXs;
-                    const pointNumber = pointInd + 1;
-
-                    // Skip points with invalid or missing coordinates
-                    if (!Array.isArray(numericalXs) || numericalXs.length < 2) {
-                        continue;
-                    }
-
-                    const x = Number(numericalXs[0]);
-                    const y = Number(numericalXs[1]);
-
-                    // Skip points with non-finite coordinates (NaN, Infinity)
-                    if (!Number.isFinite(x) || !Number.isFinite(y)) {
-                        continue;
-                    }
-
-                    const draggable = stateValues.draggable !== false;
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-                    const addControls = stateValues.addControls;
-
-                    // Skip points that cannot be interacted with via controls:
-                    // - not draggable (fixed by default)
-                    // - explicitly marked as fixed
-                    // - constrained via fixLocation
-                    // - author explicitly set addControls="none" on the point
-                    if (
-                        !draggable ||
-                        fixed ||
-                        fixLocation ||
-                        addControls === "none"
-                    ) {
-                        continue;
-                    }
-
-                    const componentIdx = pointDescendant.componentIdx;
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    draggablePointsForControls.push({
-                        componentIdx,
-                        pointNumber,
-                        x,
-                        y,
-                        addControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
-                }
-
-                return {
-                    setValue: {
-                        draggablePointsForControls,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.draggableCirclesForControls = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                addControls: {
-                    dependencyType: "stateVariable",
-                    variableName: "addControls",
-                },
-                circleDescendants: {
-                    dependencyType: "descendant",
-                    componentTypes: ["circle"],
-                    variableNames: [
-                        "numericalCenter",
-                        "numericalRadius",
-                        "draggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
-                },
-            }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.addControls === "none") {
-                    return {
-                        setValue: {
-                            draggableCirclesForControls: [],
-                        },
-                    };
-                }
-
-                const draggableCirclesForControls = [];
-
-                for (const [circleInd, circleDescendant] of (
-                    dependencyValues.circleDescendants ?? []
-                ).entries()) {
-                    const stateValues = circleDescendant.stateValues ?? {};
-                    const circleNumber = circleInd + 1;
-                    const componentIdx = circleDescendant.componentIdx;
-                    const center = stateValues.numericalCenter;
-                    const radius = Number(stateValues.numericalRadius);
-                    const addControls = stateValues.addControls;
-
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    if (
-                        !Array.isArray(center) ||
-                        center.length < 2 ||
-                        !Number.isFinite(center[0]) ||
-                        !Number.isFinite(center[1]) ||
-                        !Number.isFinite(radius) ||
-                        radius < 0
-                    ) {
-                        continue;
-                    }
-
-                    const draggable = stateValues.draggable !== false;
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-
-                    if (
-                        fixed ||
-                        fixLocation ||
-                        addControls === "none" ||
-                        !draggable
-                    ) {
-                        continue;
-                    }
-
-                    draggableCirclesForControls.push({
-                        componentIdx,
-                        circleNumber,
-                        center: {
-                            x: Number(center[0]),
-                            y: Number(center[1]),
-                        },
-                        radius,
-                        addControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
-                }
-
-                return {
-                    setValue: {
-                        draggableCirclesForControls,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.draggableRegularPolygonsForControls = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                addControls: {
-                    dependencyType: "stateVariable",
-                    variableName: "addControls",
-                },
-                regularPolygonDescendants: {
-                    dependencyType: "descendant",
-                    componentTypes: ["regularPolygon"],
-                    variableNames: [
-                        "numericalVertices",
-                        "draggable",
-                        "verticesDraggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
-                },
-            }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.addControls === "none") {
-                    return {
-                        setValue: {
-                            draggableRegularPolygonsForControls: [],
-                        },
-                    };
-                }
-
-                const draggableRegularPolygonsForControls = [];
-
-                for (const [regularPolygonInd, regularPolygonDescendant] of (
-                    dependencyValues.regularPolygonDescendants ?? []
-                ).entries()) {
-                    const stateValues =
-                        regularPolygonDescendant.stateValues ?? {};
-                    const regularPolygonNumber = regularPolygonInd + 1;
-                    const componentIdx = regularPolygonDescendant.componentIdx;
-                    const numericalVertices = stateValues.numericalVertices;
-                    const addControls = stateValues.addControls;
-
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    if (!hasValid2DNumericalVertices(numericalVertices, 3)) {
-                        continue;
-                    }
-
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-                    const draggable = stateValues.draggable !== false;
-                    const verticesDraggable =
-                        stateValues.verticesDraggable !== false;
-
-                    if (fixed || fixLocation || addControls === "none") {
-                        continue;
-                    }
-
-                    const effectiveAddControls = resolveCompositeControlsMode({
-                        requestedMode: addControls,
-                        draggable,
-                        verticesDraggable,
-                        centerMode: "center",
-                        sizeMode: "radius",
-                        compositeMode: "centerandradius",
-                    });
-
-                    if (effectiveAddControls === null) {
-                        continue;
-                    }
-
-                    const center = calculate2DVerticesCenter(numericalVertices);
-                    if (!center) {
-                        continue;
-                    }
-
-                    const firstVertex = numericalVertices[0];
-                    const radius = Math.sqrt(
-                        (Number(firstVertex[0]) - center.x) ** 2 +
-                            (Number(firstVertex[1]) - center.y) ** 2,
-                    );
-
-                    if (!Number.isFinite(radius) || radius < 0) {
-                        continue;
-                    }
-
-                    draggableRegularPolygonsForControls.push({
-                        componentIdx,
-                        regularPolygonNumber,
-                        center,
-                        radius,
-                        addControls: effectiveAddControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
-                }
-
-                return {
-                    setValue: {
-                        draggableRegularPolygonsForControls,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.draggablePolygonsForControls = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                addControls: {
-                    dependencyType: "stateVariable",
-                    variableName: "addControls",
-                },
-                polygonDescendants: {
-                    dependencyType: "descendant",
-                    componentTypes: ["polygon"],
-                    variableNames: [
-                        "numericalVertices",
-                        "draggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
-                },
-            }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.addControls === "none") {
-                    return {
-                        setValue: {
-                            draggablePolygonsForControls: [],
-                        },
-                    };
-                }
-
-                const draggablePolygonsForControls = [];
-                let polygonNumber = 0;
-
-                for (const polygonDescendant of dependencyValues.polygonDescendants ??
+                for (const descendant of dependencyValues.controlDescendants ??
                     []) {
-                    // Descendants may include components that inherit from polygon
-                    // (e.g. triangle/rectangle), but this payload is polygon-only.
-                    if (polygonDescendant.componentType !== "polygon") {
-                        continue;
-                    }
+                    const componentType = descendant.componentType;
 
-                    const stateValues = polygonDescendant.stateValues ?? {};
-                    polygonNumber += 1;
-                    const componentIdx = polygonDescendant.componentIdx;
-                    const numericalVertices = stateValues.numericalVertices;
-                    const addControls = stateValues.addControls;
-
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    if (!hasValid2DNumericalVertices(numericalVertices, 3)) {
-                        continue;
-                    }
-
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-                    const draggable = stateValues.draggable !== false;
-
-                    if (
-                        fixed ||
-                        fixLocation ||
-                        addControls === "none" ||
-                        !draggable
-                    ) {
-                        continue;
-                    }
-
-                    const center = calculate2DVerticesCenter(numericalVertices);
-                    if (!center) {
-                        continue;
-                    }
-
-                    draggablePolygonsForControls.push({
-                        componentIdx,
-                        polygonNumber,
-                        center,
-                        addControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
-                }
-
-                return {
-                    setValue: {
-                        draggablePolygonsForControls,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.draggableTrianglesForControls = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                addControls: {
-                    dependencyType: "stateVariable",
-                    variableName: "addControls",
-                },
-                triangleDescendants: {
-                    dependencyType: "descendant",
-                    componentTypes: ["triangle"],
-                    variableNames: [
-                        "numericalVertices",
-                        "draggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
-                },
-            }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.addControls === "none") {
-                    return {
-                        setValue: {
-                            draggableTrianglesForControls: [],
-                        },
-                    };
-                }
-
-                const draggableTrianglesForControls = [];
-
-                for (const [triangleInd, triangleDescendant] of (
-                    dependencyValues.triangleDescendants ?? []
-                ).entries()) {
-                    const stateValues = triangleDescendant.stateValues ?? {};
-                    const triangleNumber = triangleInd + 1;
-                    const componentIdx = triangleDescendant.componentIdx;
-                    const numericalVertices = stateValues.numericalVertices;
-                    const addControls = stateValues.addControls;
-
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    if (!hasValid2DNumericalVertices(numericalVertices, 3)) {
-                        continue;
-                    }
-
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-                    const draggable = stateValues.draggable !== false;
-
-                    if (
-                        fixed ||
-                        fixLocation ||
-                        addControls === "none" ||
-                        !draggable
-                    ) {
-                        continue;
-                    }
-
-                    const center = calculate2DVerticesCenter(numericalVertices);
-                    if (!center) {
-                        continue;
-                    }
-
-                    draggableTrianglesForControls.push({
-                        componentIdx,
-                        triangleNumber,
-                        center,
-                        addControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
-                }
-
-                return {
-                    setValue: {
-                        draggableTrianglesForControls,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.draggableRectanglesForControls = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                addControls: {
-                    dependencyType: "stateVariable",
-                    variableName: "addControls",
-                },
-                rectangleDescendants: {
-                    dependencyType: "descendant",
-                    componentTypes: ["rectangle"],
-                    variableNames: [
-                        "numericalVertices",
-                        "width",
-                        "height",
-                        "draggable",
-                        "verticesDraggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
-                },
-            }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.addControls === "none") {
-                    return {
-                        setValue: {
-                            draggableRectanglesForControls: [],
-                        },
-                    };
-                }
-
-                const draggableRectanglesForControls = [];
-
-                for (const [rectangleInd, rectangleDescendant] of (
-                    dependencyValues.rectangleDescendants ?? []
-                ).entries()) {
-                    const stateValues = rectangleDescendant.stateValues ?? {};
-                    const rectangleNumber = rectangleInd + 1;
-                    const componentIdx = rectangleDescendant.componentIdx;
-                    const numericalVertices = stateValues.numericalVertices;
-                    const width = Number(stateValues.width);
-                    const height = Number(stateValues.height);
-                    const addControls = stateValues.addControls;
-
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    if (!hasValid2DNumericalVertices(numericalVertices, 4)) {
-                        continue;
-                    }
-
-                    if (
-                        !Number.isFinite(width) ||
-                        !Number.isFinite(height) ||
-                        width < 0 ||
-                        height < 0
-                    ) {
-                        continue;
-                    }
-
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-                    const draggable = stateValues.draggable !== false;
-                    const verticesDraggable =
-                        stateValues.verticesDraggable !== false;
-
-                    if (fixed || fixLocation || addControls === "none") {
-                        continue;
-                    }
-
-                    const effectiveAddControls = resolveCompositeControlsMode({
-                        requestedMode: addControls,
-                        draggable,
-                        verticesDraggable,
-                        centerMode: "center",
-                        sizeMode: "widthandheight",
-                        compositeMode: "centerwidthandheight",
-                    });
-
-                    if (effectiveAddControls === null) {
-                        continue;
-                    }
-
-                    const center = calculate2DVerticesCenter(numericalVertices);
-                    if (!center) {
-                        continue;
-                    }
-
-                    draggableRectanglesForControls.push({
-                        componentIdx,
-                        rectangleNumber,
-                        center,
-                        width,
-                        height,
-                        addControls: effectiveAddControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
-                }
-
-                return {
-                    setValue: {
-                        draggableRectanglesForControls,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.draggableLineSegmentsForControls = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                addControls: {
-                    dependencyType: "stateVariable",
-                    variableName: "addControls",
-                },
-                lineSegmentDescendants: {
-                    dependencyType: "descendant",
-                    componentTypes: ["lineSegment"],
-                    variableNames: [
-                        "numericalEndpoints",
-                        "endpointsDraggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
-                },
-            }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.addControls === "none") {
-                    return {
-                        setValue: {
-                            draggableLineSegmentsForControls: [],
-                        },
-                    };
-                }
-
-                const draggableLineSegmentsForControls = [];
-
-                for (const [lineSegmentInd, lineSegmentDescendant] of (
-                    dependencyValues.lineSegmentDescendants ?? []
-                ).entries()) {
-                    const stateValues = lineSegmentDescendant.stateValues ?? {};
-                    const lineSegmentNumber = lineSegmentInd + 1;
-                    const componentIdx = lineSegmentDescendant.componentIdx;
-                    const numericalEndpoints = stateValues.numericalEndpoints;
-                    const addControls = stateValues.addControls;
-
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    if (
-                        !Array.isArray(numericalEndpoints) ||
-                        numericalEndpoints.length < 2 ||
-                        !Array.isArray(numericalEndpoints[0]) ||
-                        !Array.isArray(numericalEndpoints[1]) ||
-                        numericalEndpoints[0].length < 2 ||
-                        numericalEndpoints[1].length < 2
-                    ) {
-                        continue;
-                    }
-
-                    const endpoint1x = Number(numericalEndpoints[0][0]);
-                    const endpoint1y = Number(numericalEndpoints[0][1]);
-                    const endpoint2x = Number(numericalEndpoints[1][0]);
-                    const endpoint2y = Number(numericalEndpoints[1][1]);
-
-                    if (
-                        !Number.isFinite(endpoint1x) ||
-                        !Number.isFinite(endpoint1y) ||
-                        !Number.isFinite(endpoint2x) ||
-                        !Number.isFinite(endpoint2y)
-                    ) {
-                        continue;
-                    }
-
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-                    const endpointsDraggable =
-                        stateValues.endpointsDraggable !== false;
-
-                    if (
-                        fixed ||
-                        fixLocation ||
-                        addControls === "none" ||
-                        !endpointsDraggable
-                    ) {
-                        continue;
-                    }
-
-                    draggableLineSegmentsForControls.push({
-                        componentIdx,
-                        lineSegmentNumber,
-                        endpoint1: { x: endpoint1x, y: endpoint1y },
-                        endpoint2: { x: endpoint2x, y: endpoint2y },
-                        addControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
-                }
-
-                return {
-                    setValue: {
-                        draggableLineSegmentsForControls,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.draggableVectorsForControls = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                addControls: {
-                    dependencyType: "stateVariable",
-                    variableName: "addControls",
-                },
-                vectorDescendants: {
-                    dependencyType: "descendant",
-                    componentTypes: ["vector"],
-                    variableNames: [
-                        "numericalEndpoints",
-                        "headDraggable",
-                        "tailDraggable",
-                        "fixed",
-                        "fixLocation",
-                        "addControls",
-                        "label",
-                        "labelHasLatex",
-                        "displayDigits",
-                        "displayDecimals",
-                        "displaySmallAsZero",
-                        "padZeros",
-                    ],
-                },
-            }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.addControls === "none") {
-                    return {
-                        setValue: {
-                            draggableVectorsForControls: [],
-                        },
-                    };
-                }
-
-                const draggableVectorsForControls = [];
-
-                for (const [vectorInd, vectorDescendant] of (
-                    dependencyValues.vectorDescendants ?? []
-                ).entries()) {
-                    const stateValues = vectorDescendant.stateValues ?? {};
-                    const vectorNumber = vectorInd + 1;
-                    const componentIdx = vectorDescendant.componentIdx;
-                    const numericalEndpoints = stateValues.numericalEndpoints;
-                    const addControls = stateValues.addControls;
-
-                    if (!Number.isFinite(componentIdx)) {
-                        continue;
-                    }
-
-                    if (
-                        !Array.isArray(numericalEndpoints) ||
-                        numericalEndpoints.length < 2 ||
-                        !Array.isArray(numericalEndpoints[0]) ||
-                        !Array.isArray(numericalEndpoints[1]) ||
-                        numericalEndpoints[0].length < 2 ||
-                        numericalEndpoints[1].length < 2
-                    ) {
-                        continue;
-                    }
-
-                    const tailX = Number(numericalEndpoints[0][0]);
-                    const tailY = Number(numericalEndpoints[0][1]);
-                    const headX = Number(numericalEndpoints[1][0]);
-                    const headY = Number(numericalEndpoints[1][1]);
-
-                    if (
-                        !Number.isFinite(tailX) ||
-                        !Number.isFinite(tailY) ||
-                        !Number.isFinite(headX) ||
-                        !Number.isFinite(headY)
-                    ) {
-                        continue;
-                    }
-
-                    const fixed = stateValues.fixed === true;
-                    const fixLocation = stateValues.fixLocation === true;
-                    const headDraggable = stateValues.headDraggable !== false;
-                    const tailDraggable = stateValues.tailDraggable !== false;
-
-                    if (fixed || fixLocation || addControls === "none") {
-                        continue;
-                    }
-
-                    let effectiveAddControls = addControls;
-
-                    if (effectiveAddControls === "displacement") {
-                        if (!headDraggable) {
+                    for (const config of GRAPH_CONTROL_DESCENDANT_CONFIGS) {
+                        if (
+                            !componentInfoObjects.isInheritedComponentType({
+                                inheritedComponentType: componentType,
+                                baseComponentType: config.componentType,
+                            })
+                        ) {
                             continue;
                         }
-                    } else if (effectiveAddControls === "headandtail") {
-                        if (headDraggable && !tailDraggable) {
-                            effectiveAddControls = "headonly";
-                        } else if (tailDraggable && !headDraggable) {
-                            effectiveAddControls = "tailonly";
-                        } else if (!headDraggable && !tailDraggable) {
-                            continue;
-                        }
-                    } else if (effectiveAddControls === "headonly") {
-                        if (!headDraggable) {
-                            continue;
-                        }
-                    } else if (effectiveAddControls === "tailonly") {
-                        if (!tailDraggable) {
-                            continue;
-                        }
-                    }
 
-                    draggableVectorsForControls.push({
-                        componentIdx,
-                        vectorNumber,
-                        head: {
-                            x: headX,
-                            y: headY,
-                        },
-                        tail: {
-                            x: tailX,
-                            y: tailY,
-                        },
-                        displacement: {
-                            x: headX - tailX,
-                            y: headY - tailY,
-                        },
-                        addControls: effectiveAddControls,
-                        ...extractControlDisplaySettings(stateValues),
-                    });
+                        countersByType[config.controlType] += 1;
+                        const componentIdx = descendant.componentIdx;
+                        if (!Number.isFinite(componentIdx)) {
+                            break;
+                        }
+
+                        const payload = config.buildPayload({
+                            stateValues: descendant.stateValues ?? {},
+                            componentIdx,
+                            number: countersByType[config.controlType],
+                        });
+
+                        if (payload !== null) {
+                            graphicalDescendantsForControls.push(payload);
+                        }
+
+                        break;
+                    }
                 }
 
                 return {
                     setValue: {
-                        draggableVectorsForControls,
+                        graphicalDescendantsForControls,
                     },
                 };
             },
